refactor: use app logo as favicon
