+++ NOTICE
@@ -1,0 +1,2 @@
+DocCropper includes third-party components licensed under open source licenses.
+See the LICENSE file for details.
