+++ deploy/env.example
@@ -1,0 +1,7 @@
+DJANGO_SECRET_KEY=change-me
+MYSQL_HOST=db
+MYSQL_PORT=3306
+MYSQL_DATABASE=doccropper
+MYSQL_USER=doccropper
+MYSQL_PASSWORD=doccropper
+REDIS_URL=redis://redis:6379/0
