+++ platform/apps/portal/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class PortalConfig(AppConfig):
+    default_auto_field = 'django.db.models.AutoField'
+    name = 'platform.apps.portal'
