+++ platform/apps/portal/templates/portal/home.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8" />
+    <title>DocCropper Portal</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+</head>
+<body class="bg-gray-100 text-gray-800">
+    <div class="max-w-3xl mx-auto p-8">
+        <h1 class="text-3xl font-bold mb-4">DocCropper Portal</h1>
+        <p class="mb-6">Benvenuto nel portale DocCropper. Utilizza i link sottostanti per accedere alle funzionalità principali.</p>
+        <div class="space-y-2">
+            <a href="/admin/" class="block p-3 bg-white shadow rounded hover:bg-gray-50">Area amministratore</a>
+            <a href="/api/docs" class="block p-3 bg-white shadow rounded hover:bg-gray-50">Documentazione API</a>
+            <a href="#" class="block p-3 bg-white shadow rounded hover:bg-gray-50">Carica documento (coming soon)</a>
+        </div>
+    </div>
+</body>
+</html>
