+++ platform/apps/portal/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('', views.home, name='portal-home'),
+]
