+++ platform/apps/portal/views.py
@@ -1,0 +1,6 @@
+from django.shortcuts import render
+
+
+def home(request):
+    """Render the main portal page using the Expressive template."""
+    return render(request, 'expressive/index.html')
