+++ platform/config/celery.py
@@ -1,0 +1,8 @@
+import os
+from celery import Celery
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'platform.config.settings')
+
+app = Celery('platform')
+app.config_from_object('django.conf:settings', namespace='CELERY')
+app.autodiscover_tasks()
