+++ plugin/README.md
@@ -1,0 +1,6 @@
+# Plugin Directory
+
+- `core/`: open source plugins released under Apache-2.0.
+- `pro/`: proprietary plugins available under a separate commercial EULA.
+
+Place community plugins in `core/` and commercial additions in `pro/`.
