+++ plugin/core/colormode/__init__.py
