+++ plugin/core/colormode/settings.json
