+++ plugin/core/compresspdf/__init__.py
