+++ plugin/core/crop/README.md
