+++ plugin/core/docuseal/settings.json
