+++ plugin/core/downloadpng/settings.json
