+++ plugin/core/imageeditor/__init__.py
