+++ plugin/core/imageeditor/settings.json
