+++ plugin/core/pageselect/settings.json
