+++ plugin/core/remotesign/settings.json
