+++ plugin/core/watermark/__init__.py
