+++ plugin/core/watermark/settings.json
