+++ plugin/pro/EULA.md
@@ -1,0 +1,3 @@
+# DocCropper Pro EULA
+
+This is a placeholder for the proprietary license agreement.
