+++ requirements.txt
@@ -8,7 +8,6 @@ rembg
 google-auth
 
 pystray
-python-dotenv
 pyhanko
 pytesseract
 PyMuPDF
@@ -22,3 +21,15 @@ aiosqlite
 passlib[bcrypt]
 python-jose
 email-validator
+django>=5.0
+djangorestframework
+django-allauth
+django-guardian
+django-tailwind
+mysqlclient
+uvicorn[standard]
+celery
+redis
+python-dotenv
+bcrypt==4.1.2
+setproctitle
