+++ scripts/start_DocCropper.bat
@@ -59,7 +59,7 @@ if exist "!PID_FILE!" (
 
 if "!TRAY_RUNNING!"=="0" (
     echo [INFO] Avvio tray helper >> "!LOG_FILE!"
-    start "" pythonw doccropper_tray.pyw >> "!LOG_FILE!" 2>&1
+    start "" cmd /c "set DOCROPPER_PROC=DocCropperTray && pythonw doccropper_tray.pyw" >> "!LOG_FILE!" 2>&1
     timeout /t 2 >nul
 )
 
@@ -71,7 +71,7 @@ if exist "!PID_FILE!" (
 )
 
 set "OPEN_URL=%DOCROPPER_OPEN_URL%"
-if "%OPEN_URL%"=="" set "OPEN_URL=http://localhost:%PORT%"
+if "%OPEN_URL%"=="" set "OPEN_URL=http://localhost:%PORT%/"
 
 if "!SERVER_RUNNING!"=="1" (
     echo [INFO] DocCropper gia in esecuzione con PID !PID! >> "!LOG_FILE!"
@@ -106,7 +106,7 @@ python main.py --stop >> "!LOG_FILE!" 2>&1
 
 :: Launch application
 echo [INFO] Avvio DocCropper sulla porta %PORT% >> "!LOG_FILE!"
-start "" /b python main.py --port %PORT% >> "!LOG_FILE!" 2>&1
+start "" /b cmd /c "set DOCROPPER_PROC=DocCropper && python main.py --port %PORT%" >> "!LOG_FILE!" 2>&1
 if "%DOCROPPER_TUNNEL%"=="true" (
     where cloudflared >nul 2>&1 && (
         echo Starting Cloudflare Tunnel... >> "!LOG_FILE!"
